refactor: remove `ChainPipeline` and related functionality from stream module
